Remove all recipe advancements that are merely mark as impossible
The set of impossible recipe advancements is inconsistent to say the least. It's also not clear why a SkyBlock map would
care to remove these advancements. Perhaps the assumption is that only experienced players will play it and so don't
need these advancements? If in the future we decide to go that route, we should use an automated process to exclude all
recipe advancements based on the latest server JAR.